trabalhando com while

+++ 2-2_-_repeticao_for/exec055_-_maior_e_menor_sequencia.py
@@ -5,7 +5,7 @@ leia o peso de cinco pessoas e descubra qual o mais pesado e o leve
 maior = 0
 menor = 0
 
-quant_numeros = float(input("Quantos números irá informar? "))
+quant_numeros = int(input("Quantos números irá informar? "))
 
 for c in range (0,quant_numeros):
     num = float(input(f'Informe o peso da pessao nº {c+1}º: '))
